Added freeimage support

+++ TestGlfw/Font.cpp
@@ -82,7 +82,7 @@ void Font::Load(const std::string& ttfPath, int height)
 			texture,
 			glm::ivec2(m_Face->glyph->bitmap.width, m_Face->glyph->bitmap.rows),
 			glm::ivec2(m_Face->glyph->bitmap_left, m_Face->glyph->bitmap_top),
-			m_Face->glyph->advance.x
+			static_cast<GLuint>(m_Face->glyph->advance.x)
 		};
 		m_Characters.insert(std::pair<GLchar, Character>(c, character));
 	}
